Edit BUG, Typo, etc..
Edit BUG, Typo, etc.. (Student_StudyStatusService)

+++ src/main/java/LikeLion11th/YouQuiz_Project/study/service/StudyService.java
@@ -64,7 +64,7 @@ public class StudyService {
 
     }
 
-    public ChapterDto readChapter(int chap_id){
+    public ChapterDto readChapter(long chap_id){
 
         ChapterDto chapterDto = new ChapterDto();
